增加检查是否有新消息的接口

+++ library/algorithm/snowFlake/snowFlake.go
@@ -71,6 +71,6 @@ func (w *Worker) GetId() int64 {
 
 	// 第一段 now - epoch 为该算法目前已经奔跑了xxx毫秒
 	// 如果在程序跑了一段时间修改了epoch这个值 可能会导致生成相同的ID
-	ID := int64((now-epoch)<<timeShift | (w.workerId << workerShift) | (w.number))
+	ID := (now-epoch)<<timeShift | (w.workerId << workerShift) | (w.number)
 	return ID
 }
